Better player service and fix #998
