+++ app/src/main/java/com/zionhuang/music/constants/Dimensions.kt
@@ -26,4 +26,6 @@ val AlbumThumbnailSize = 144.dp
 
 val ThumbnailCornerRadius = 6.dp
 
+val PlayerHorizontalPadding = 32.dp
+
 val NavigationBarAnimationSpec = spring<Dp>(stiffness = Spring.StiffnessMediumLow)
