+++ app/src/main/java/com/zionhuang/music/playback/MediaLibrarySessionCallback.kt
@@ -1,0 +1,300 @@
+package com.zionhuang.music.playback
+
+import android.content.ContentResolver
+import android.content.Context
+import android.net.Uri
+import android.os.Bundle
+import androidx.annotation.DrawableRes
+import androidx.core.net.toUri
+import androidx.media3.common.MediaItem
+import androidx.media3.common.MediaMetadata
+import androidx.media3.exoplayer.offline.Download
+import androidx.media3.session.LibraryResult
+import androidx.media3.session.MediaLibraryService
+import androidx.media3.session.MediaLibraryService.MediaLibrarySession
+import androidx.media3.session.MediaSession
+import androidx.media3.session.SessionCommand
+import androidx.media3.session.SessionResult
+import com.google.common.collect.ImmutableList
+import com.google.common.util.concurrent.Futures
+import com.google.common.util.concurrent.ListenableFuture
+import com.zionhuang.music.R
+import com.zionhuang.music.constants.MediaSessionConstants
+import com.zionhuang.music.constants.SongSortType
+import com.zionhuang.music.db.MusicDatabase
+import com.zionhuang.music.db.entities.PlaylistEntity
+import com.zionhuang.music.db.entities.Song
+import com.zionhuang.music.extensions.toMediaItem
+import dagger.hilt.android.qualifiers.ApplicationContext
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.guava.future
+import kotlinx.coroutines.plus
+import javax.inject.Inject
+
+class MediaLibrarySessionCallback @Inject constructor(
+    @ApplicationContext val context: Context,
+    val database: MusicDatabase,
+    val downloadUtil: DownloadUtil,
+) : MediaLibrarySession.Callback {
+    private val scope = CoroutineScope(Dispatchers.Main) + Job()
+    var toggleLike: () -> Unit = {}
+    var toggleLibrary: () -> Unit = {}
+
+    override fun onConnect(session: MediaSession, controller: MediaSession.ControllerInfo): MediaSession.ConnectionResult {
+        val connectionResult = super.onConnect(session, controller)
+        return MediaSession.ConnectionResult.accept(
+            connectionResult.availableSessionCommands.buildUpon()
+                .add(MediaSessionConstants.CommandToggleLibrary)
+                .add(MediaSessionConstants.CommandToggleLike).build(),
+            connectionResult.availablePlayerCommands
+        )
+    }
+
+    override fun onCustomCommand(
+        session: MediaSession,
+        controller: MediaSession.ControllerInfo,
+        customCommand: SessionCommand,
+        args: Bundle,
+    ): ListenableFuture<SessionResult> {
+        when (customCommand.customAction) {
+            MediaSessionConstants.ACTION_TOGGLE_LIKE -> toggleLike()
+            MediaSessionConstants.ACTION_TOGGLE_LIBRARY -> toggleLibrary()
+        }
+        return Futures.immediateFuture(SessionResult(SessionResult.RESULT_SUCCESS))
+    }
+
+    override fun onGetLibraryRoot(
+        session: MediaLibrarySession,
+        browser: MediaSession.ControllerInfo,
+        params: MediaLibraryService.LibraryParams?,
+    ): ListenableFuture<LibraryResult<MediaItem>> = Futures.immediateFuture(
+        LibraryResult.ofItem(
+            MediaItem.Builder()
+                .setMediaId(MusicService.ROOT)
+                .setMediaMetadata(
+                    MediaMetadata.Builder()
+                        .setIsPlayable(false)
+                        .setIsBrowsable(false)
+                        .setMediaType(MediaMetadata.MEDIA_TYPE_FOLDER_MIXED)
+                        .build()
+                )
+                .build(),
+            params
+        )
+    )
+
+    override fun onGetChildren(
+        session: MediaLibrarySession,
+        browser: MediaSession.ControllerInfo,
+        parentId: String,
+        page: Int,
+        pageSize: Int,
+        params: MediaLibraryService.LibraryParams?,
+    ): ListenableFuture<LibraryResult<ImmutableList<MediaItem>>> = scope.future(Dispatchers.IO) {
+        LibraryResult.ofItemList(
+            when (parentId) {
+                MusicService.ROOT -> listOf(
+                    browsableMediaItem(MusicService.SONG, context.getString(R.string.songs), null, drawableUri(R.drawable.music_note), MediaMetadata.MEDIA_TYPE_PLAYLIST),
+                    browsableMediaItem(MusicService.ARTIST, context.getString(R.string.artists), null, drawableUri(R.drawable.artist), MediaMetadata.MEDIA_TYPE_FOLDER_ARTISTS),
+                    browsableMediaItem(MusicService.ALBUM, context.getString(R.string.albums), null, drawableUri(R.drawable.album), MediaMetadata.MEDIA_TYPE_FOLDER_ALBUMS),
+                    browsableMediaItem(MusicService.PLAYLIST, context.getString(R.string.playlists), null, drawableUri(R.drawable.queue_music), MediaMetadata.MEDIA_TYPE_FOLDER_PLAYLISTS)
+                )
+
+                MusicService.SONG -> database.songsByCreateDateAsc().first().map { it.toMediaItem(parentId) }
+                MusicService.ARTIST -> database.artistsByCreateDateAsc().first().map { artist ->
+                    browsableMediaItem("${MusicService.ARTIST}/${artist.id}", artist.artist.name, context.resources.getQuantityString(R.plurals.n_song, artist.songCount, artist.songCount), artist.artist.thumbnailUrl?.toUri(), MediaMetadata.MEDIA_TYPE_ARTIST)
+                }
+
+                MusicService.ALBUM -> database.albumsByCreateDateAsc().first().map { album ->
+                    browsableMediaItem("${MusicService.ALBUM}/${album.id}", album.album.title, album.artists.joinToString(), album.album.thumbnailUrl?.toUri(), MediaMetadata.MEDIA_TYPE_ALBUM)
+                }
+
+                MusicService.PLAYLIST -> {
+                    val likedSongCount = database.likedSongsCount().first()
+                    val downloadedSongCount = downloadUtil.downloads.value.size
+                    listOf(
+                        browsableMediaItem("${MusicService.PLAYLIST}/${PlaylistEntity.LIKED_PLAYLIST_ID}", context.getString(R.string.liked_songs), context.resources.getQuantityString(R.plurals.n_song, likedSongCount, likedSongCount), drawableUri(R.drawable.favorite), MediaMetadata.MEDIA_TYPE_PLAYLIST),
+                        browsableMediaItem("${MusicService.PLAYLIST}/${PlaylistEntity.DOWNLOADED_PLAYLIST_ID}", context.getString(R.string.downloaded_songs), context.resources.getQuantityString(R.plurals.n_song, downloadedSongCount, downloadedSongCount), drawableUri(R.drawable.download), MediaMetadata.MEDIA_TYPE_PLAYLIST)
+                    ) + database.playlistsByCreateDateAsc().first().map { playlist ->
+                        browsableMediaItem("${MusicService.PLAYLIST}/${playlist.id}", playlist.playlist.name, context.resources.getQuantityString(R.plurals.n_song, playlist.songCount, playlist.songCount), playlist.thumbnails.firstOrNull()?.toUri(), MediaMetadata.MEDIA_TYPE_PLAYLIST)
+                    }
+                }
+
+                else -> when {
+                    parentId.startsWith("${MusicService.ARTIST}/") ->
+                        database.artistSongsByCreateDateAsc(parentId.removePrefix("${MusicService.ARTIST}/")).first().map {
+                            it.toMediaItem(parentId)
+                        }
+
+                    parentId.startsWith("${MusicService.ALBUM}/") ->
+                        database.albumSongs(parentId.removePrefix("${MusicService.ALBUM}/")).first().map {
+                            it.toMediaItem(parentId)
+                        }
+
+                    parentId.startsWith("${MusicService.PLAYLIST}/") ->
+                        when (val playlistId = parentId.removePrefix("${MusicService.PLAYLIST}/")) {
+                            PlaylistEntity.LIKED_PLAYLIST_ID -> database.likedSongs(SongSortType.CREATE_DATE, true)
+                            PlaylistEntity.DOWNLOADED_PLAYLIST_ID -> {
+                                val downloads = downloadUtil.downloads.value
+                                database.allSongs()
+                                    .flowOn(Dispatchers.IO)
+                                    .map { songs ->
+                                        songs.filter {
+                                            downloads[it.id]?.state == Download.STATE_COMPLETED
+                                        }
+                                    }
+                                    .map { songs ->
+                                        songs.map { it to downloads[it.id] }
+                                            .sortedBy { it.second?.updateTimeMs ?: 0L }
+                                            .map { it.first }
+                                    }
+                            }
+
+                            else -> database.playlistSongs(playlistId).map { list ->
+                                list.map { it.song }
+                            }
+                        }.first().map {
+                            it.toMediaItem(parentId)
+                        }
+
+                    else -> emptyList()
+                }
+            },
+            params
+        )
+    }
+
+    override fun onGetItem(
+        session: MediaLibrarySession,
+        browser: MediaSession.ControllerInfo,
+        mediaId: String,
+    ): ListenableFuture<LibraryResult<MediaItem>> = scope.future(Dispatchers.IO) {
+        database.song(mediaId).first()?.toMediaItem()?.let {
+            LibraryResult.ofItem(it, null)
+        } ?: LibraryResult.ofError(LibraryResult.RESULT_ERROR_UNKNOWN)
+    }
+
+    override fun onSetMediaItems(
+        mediaSession: MediaSession,
+        controller: MediaSession.ControllerInfo,
+        mediaItems: MutableList<MediaItem>,
+        startIndex: Int,
+        startPositionMs: Long,
+    ): ListenableFuture<MediaSession.MediaItemsWithStartPosition> = scope.future {
+        // Play from Android Auto
+        val defaultResult = MediaSession.MediaItemsWithStartPosition(emptyList(), startIndex, startPositionMs)
+        val path = mediaItems.firstOrNull()?.mediaId?.split("/")
+            ?: return@future defaultResult
+        when (path.firstOrNull()) {
+            MusicService.SONG -> {
+                val songId = path.getOrNull(1) ?: return@future defaultResult
+                val allSongs = database.songsByCreateDateAsc().first()
+                MediaSession.MediaItemsWithStartPosition(
+                    allSongs.map { it.toMediaItem() },
+                    allSongs.indexOfFirst { it.id == songId }.takeIf { it != -1 } ?: 0,
+                    startPositionMs
+                )
+            }
+
+            MusicService.ARTIST -> {
+                val songId = path.getOrNull(2) ?: return@future defaultResult
+                val artistId = path.getOrNull(1) ?: return@future defaultResult
+                val songs = database.artistSongsByCreateDateAsc(artistId).first()
+                MediaSession.MediaItemsWithStartPosition(
+                    songs.map { it.toMediaItem() },
+                    songs.indexOfFirst { it.id == songId }.takeIf { it != -1 } ?: 0,
+                    startPositionMs
+                )
+            }
+
+            MusicService.ALBUM -> {
+                val songId = path.getOrNull(2) ?: return@future defaultResult
+                val albumId = path.getOrNull(1) ?: return@future defaultResult
+                val albumWithSongs = database.albumWithSongs(albumId).first() ?: return@future defaultResult
+                MediaSession.MediaItemsWithStartPosition(
+                    albumWithSongs.songs.map { it.toMediaItem() },
+                    albumWithSongs.songs.indexOfFirst { it.id == songId }.takeIf { it != -1 } ?: 0,
+                    startPositionMs
+                )
+            }
+
+            MusicService.PLAYLIST -> {
+                val songId = path.getOrNull(2) ?: return@future defaultResult
+                val playlistId = path.getOrNull(1) ?: return@future defaultResult
+                val songs = when (playlistId) {
+                    PlaylistEntity.LIKED_PLAYLIST_ID -> database.likedSongs(SongSortType.CREATE_DATE, descending = true)
+                    PlaylistEntity.DOWNLOADED_PLAYLIST_ID -> {
+                        val downloads = downloadUtil.downloads.value
+                        database.allSongs()
+                            .flowOn(Dispatchers.IO)
+                            .map { songs ->
+                                songs.filter {
+                                    downloads[it.id]?.state == Download.STATE_COMPLETED
+                                }
+                            }
+                            .map { songs ->
+                                songs.map { it to downloads[it.id] }
+                                    .sortedBy { it.second?.updateTimeMs ?: 0L }
+                                    .map { it.first }
+                            }
+                    }
+
+                    else -> database.playlistSongs(playlistId).map { list ->
+                        list.map { it.song }
+                    }
+                }.first()
+                MediaSession.MediaItemsWithStartPosition(
+                    songs.map { it.toMediaItem() },
+                    songs.indexOfFirst { it.id == songId }.takeIf { it != -1 } ?: 0,
+                    startPositionMs
+                )
+            }
+
+            else -> defaultResult
+        }
+    }
+
+    private fun drawableUri(@DrawableRes id: Int) = Uri.Builder()
+        .scheme(ContentResolver.SCHEME_ANDROID_RESOURCE)
+        .authority(context.resources.getResourcePackageName(id))
+        .appendPath(context.resources.getResourceTypeName(id))
+        .appendPath(context.resources.getResourceEntryName(id))
+        .build()
+
+    private fun browsableMediaItem(id: String, title: String, subtitle: String?, iconUri: Uri?, mediaType: Int = MediaMetadata.MEDIA_TYPE_MUSIC) =
+        MediaItem.Builder()
+            .setMediaId(id)
+            .setMediaMetadata(
+                MediaMetadata.Builder()
+                    .setTitle(title)
+                    .setSubtitle(subtitle)
+                    .setArtist(subtitle)
+                    .setArtworkUri(iconUri)
+                    .setIsPlayable(false)
+                    .setIsBrowsable(true)
+                    .setMediaType(mediaType)
+                    .build()
+            )
+            .build()
+
+    private fun Song.toMediaItem(path: String) =
+        MediaItem.Builder()
+            .setMediaId("$path/$id")
+            .setMediaMetadata(
+                MediaMetadata.Builder()
+                    .setTitle(song.title)
+                    .setSubtitle(artists.joinToString { it.name })
+                    .setArtist(artists.joinToString { it.name })
+                    .setArtworkUri(song.thumbnailUrl?.toUri())
+                    .setIsPlayable(true)
+                    .setIsBrowsable(false)
+                    .setMediaType(MediaMetadata.MEDIA_TYPE_MUSIC)
+                    .build()
+            )
+            .build()
+}
