+++ app/src/main/java/com/zionhuang/music/playback/SleepTimer.kt
@@ -1,0 +1,61 @@
+package com.zionhuang.music.playback
+
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import androidx.media3.common.MediaItem
+import androidx.media3.common.Player
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+import kotlin.time.Duration.Companion.minutes
+
+class SleepTimer(
+    private val scope: CoroutineScope,
+    val player: Player,
+) : Player.Listener {
+    private var sleepTimerJob: Job? = null
+    var triggerTime by mutableStateOf(-1L)
+        private set
+    var pauseWhenSongEnd by mutableStateOf(false)
+        private set
+    val isActive: Boolean
+        get() = triggerTime != -1L || pauseWhenSongEnd
+
+    fun start(minute: Int) {
+        sleepTimerJob?.cancel()
+        sleepTimerJob = null
+        if (minute == -1) {
+            pauseWhenSongEnd = true
+        } else {
+            triggerTime = System.currentTimeMillis() + minute.minutes.inWholeMilliseconds
+            sleepTimerJob = scope.launch {
+                delay(minute.minutes)
+                player.pause()
+                triggerTime = -1L
+            }
+        }
+    }
+
+    fun clear() {
+        sleepTimerJob?.cancel()
+        sleepTimerJob = null
+        pauseWhenSongEnd = false
+        triggerTime = -1L
+    }
+
+    override fun onMediaItemTransition(mediaItem: MediaItem?, reason: Int) {
+        if (pauseWhenSongEnd) {
+            pauseWhenSongEnd = false
+            player.pause()
+        }
+    }
+
+    override fun onPlaybackStateChanged(@Player.State playbackState: Int) {
+        if (playbackState == Player.STATE_ENDED && pauseWhenSongEnd) {
+            pauseWhenSongEnd = false
+            player.pause()
+        }
+    }
+}
