+++ app/src/main/java/com/zionhuang/music/ui/screens/YouTubeBrowseScreen.kt
@@ -94,26 +94,22 @@ fun YouTubeBrowseScreen(
                                         is SongItem -> YouTubeSongMenu(
                                             song = item,
                                             navController = navController,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is AlbumItem -> YouTubeAlbumMenu(
                                             albumItem = item,
                                             navController = navController,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is ArtistItem -> YouTubeArtistMenu(
                                             artist = item,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is PlaylistItem -> YouTubePlaylistMenu(
                                             playlist = item,
-                                            playerConnection = playerConnection,
                                             coroutineScope = coroutineScope,
                                             onDismiss = menuState::dismiss
                                         )
