+++ app/src/main/java/com/zionhuang/music/ui/screens/artist/ArtistItemsScreen.kt
@@ -113,26 +113,22 @@ fun ArtistItemsScreen(
                                         is SongItem -> YouTubeSongMenu(
                                             song = item,
                                             navController = navController,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is AlbumItem -> YouTubeAlbumMenu(
                                             albumItem = item,
                                             navController = navController,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is ArtistItem -> YouTubeArtistMenu(
                                             artist = item,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is PlaylistItem -> YouTubePlaylistMenu(
                                             playlist = item,
-                                            playerConnection = playerConnection,
                                             coroutineScope = coroutineScope,
                                             onDismiss = menuState::dismiss
                                         )
@@ -210,26 +206,22 @@ fun ArtistItemsScreen(
                                         is SongItem -> YouTubeSongMenu(
                                             song = item,
                                             navController = navController,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is AlbumItem -> YouTubeAlbumMenu(
                                             albumItem = item,
                                             navController = navController,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is ArtistItem -> YouTubeArtistMenu(
                                             artist = item,
-                                            playerConnection = playerConnection,
                                             onDismiss = menuState::dismiss
                                         )
 
                                         is PlaylistItem -> YouTubePlaylistMenu(
                                             playlist = item,
-                                            playerConnection = playerConnection,
                                             coroutineScope = coroutineScope,
                                             onDismiss = menuState::dismiss
                                         )
