+++ app/src/main/java/com/zionhuang/music/ui/screens/artist/ArtistScreen.kt
@@ -230,7 +230,6 @@ fun ArtistScreen(
                                             SongMenu(
                                                 originalSong = song,
                                                 navController = navController,
-                                                playerConnection = playerConnection,
                                                 onDismiss = menuState::dismiss
                                             )
                                         }
@@ -284,7 +283,6 @@ fun ArtistScreen(
                                                 YouTubeSongMenu(
                                                     song = song,
                                                     navController = navController,
-                                                    playerConnection = playerConnection,
                                                     onDismiss = menuState::dismiss
                                                 )
                                             }
@@ -339,26 +337,22 @@ fun ArtistScreen(
                                                             is SongItem -> YouTubeSongMenu(
                                                                 song = item,
                                                                 navController = navController,
-                                                                playerConnection = playerConnection,
                                                                 onDismiss = menuState::dismiss
                                                             )
 
                                                             is AlbumItem -> YouTubeAlbumMenu(
                                                                 albumItem = item,
                                                                 navController = navController,
-                                                                playerConnection = playerConnection,
                                                                 onDismiss = menuState::dismiss
                                                             )
 
                                                             is ArtistItem -> YouTubeArtistMenu(
                                                                 artist = item,
-                                                                playerConnection = playerConnection,
                                                                 onDismiss = menuState::dismiss
                                                             )
 
                                                             is PlaylistItem -> YouTubePlaylistMenu(
                                                                 playlist = item,
-                                                                playerConnection = playerConnection,
                                                                 coroutineScope = coroutineScope,
                                                                 onDismiss = menuState::dismiss
                                                             )
