+++ app/src/main/java/com/zionhuang/music/ui/screens/artist/ArtistSongsScreen.kt
@@ -124,7 +124,6 @@ fun ArtistSongsScreen(
                                     SongMenu(
                                         originalSong = song,
                                         navController = navController,
-                                        playerConnection = playerConnection,
                                         onDismiss = menuState::dismiss
                                     )
                                 }
