+++ app/src/main/java/com/zionhuang/music/ui/screens/playlist/LocalPlaylistScreen.kt
@@ -529,7 +529,6 @@ fun LocalPlaylistScreen(
                                             SongMenu(
                                                 originalSong = song.song,
                                                 navController = navController,
-                                                playerConnection = playerConnection,
                                                 onDismiss = menuState::dismiss
                                             )
                                         }
