+++ app/src/main/java/com/zionhuang/music/ui/screens/search/LocalSearchScreen.kt
@@ -154,8 +154,7 @@ fun LocalSearchScreen(
                                         menuState.show {
                                             SongMenu(
                                                 originalSong = item,
-                                                navController = navController,
-                                                playerConnection = playerConnection
+                                                navController = navController
                                             ) {
                                                 onDismiss()
                                                 menuState.dismiss()
