+++ gradle/libs.versions.toml
@@ -5,7 +5,7 @@ compose-compiler = "1.4.0"
 compose = "1.3.3"
 lifecycle = "2.6.1"
 material3 = "1.1.0-alpha05"
-media3 = "1.0.2"
+media3 = "1.1.1"
 room = "2.5.2"
 hilt = "2.46.1"
 ktor = "2.2.2"
@@ -40,7 +40,7 @@ material3-windowsize = { group = "androidx.compose.material3", name = "material3
 
 accompanist-swiperefresh = { group = "com.google.accompanist", name = "accompanist-swiperefresh", version = "0.28.0" }
 
-coil = { group = "io.coil-kt", name = "coil-compose", version = "2.2.2" }
+coil = { group = "io.coil-kt", name = "coil-compose", version = "2.3.0" }
 
 shimmer = { group = "com.valentinilk.shimmer", name = "compose-shimmer", version = "1.0.3" }
 
@@ -76,10 +76,10 @@ junit = { group = "junit", name = "junit", version = "4.13.2" }
 timber = { group = "com.jakewharton.timber", name = "timber", version = "4.7.1" }
 
 google-services = { module = "com.google.gms:google-services", version = "4.3.15" }
-firebase-bom = { group = "com.google.firebase", name = "firebase-bom", version = "32.2.0" }
+firebase-bom = { group = "com.google.firebase", name = "firebase-bom", version = "32.2.3" }
 firebase-analytics = { group = "com.google.firebase", name = "firebase-analytics-ktx" }
 firebase-crashlytics = { group = "com.google.firebase", name = "firebase-crashlytics-ktx" }
-firebase-crashlytics-plugin = { module = "com.google.firebase:firebase-crashlytics-gradle", version = "2.9.7" }
+firebase-crashlytics-plugin = { module = "com.google.firebase:firebase-crashlytics-gradle", version = "2.9.9" }
 firebase-config = { group = "com.google.firebase", name = "firebase-config-ktx" }
 firebase-perf = { group = "com.google.firebase", name = "firebase-perf-ktx" }
 firebase-perf-plugin = { module = "com.google.firebase:perf-plugin", version = "1.4.2" }
